backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.16
djangorestframework==3.15.2
idna==3.10
numpy==2.0.2
pandas==2.2.3
polib==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
django-cors-headers
python-dotenv==1.1.0
pydantic==2.11.3
requests-toolbelt==1.0.0
celery>=5.3.0
redis>=4.5.0
reportlab>=4.0.0
Pillow>=10.0.0